chore: move form-accessRights component
