+++ src/components/form-accessRights/configured-form-accessRights.jsx
@@ -3,7 +3,7 @@ import _throttle from 'lodash/throttle';
 
 import FormAccessRights from './form-accessRights.component';
 import validate from './form-accessRights.validations';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'accessRights',
